requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
geopy>=2.3.0
xmltodict>=0.13.0
matplotlib>=3.7.2
aiohttp>=3.8.4
asyncio>=3.4.3